Bring preference opt dataset on par with other datasets
