+++ src/fairseq2/datasets/preference.py
@@ -1,3 +1,9 @@
+# Copyright (c) Meta Platforms, Inc. and affiliates.
+# All rights reserved.
+#
+# This source code is licensed under the BSD-style license found in the
+# LICENSE file in the root directory of this source tree.
+
 from __future__ import annotations
 
 import json
@@ -5,7 +11,7 @@ from abc import ABC, abstractmethod
 from collections.abc import Sequence
 from dataclasses import dataclass
 from pathlib import Path
-from typing import Any, NoReturn, cast, final
+from typing import Any, cast, final
 
 import torch
 from typing_extensions import override
@@ -23,8 +29,8 @@ from fairseq2.data import (
 from fairseq2.data.text import TextTokenizer
 from fairseq2.datasets.batching import LengthBatching, StaticBatching
 from fairseq2.datasets.data_reader import DataPipelineReader
-from fairseq2.datasets.error import DatasetError
 from fairseq2.datasets.loader import AbstractDatasetLoader, DelegatingDatasetLoader
+from fairseq2.datasets.utils import _load_files_and_weights
 from fairseq2.gang import Gang
 from fairseq2.models.sequence import SequenceBatch
 from fairseq2.nn.padding import get_seqs_and_padding_mask
@@ -32,14 +38,14 @@ from fairseq2.nn.padding import get_seqs_and_padding_mask
 
 @dataclass
 class PreferenceOptimizationBatch:
-    """Represents a preference optimization batch."""
+    """Represents a preference optimization dataset batch."""
 
     chosen: SequenceBatch
     rejected: SequenceBatch
 
 
 class PreferenceOptimizationDataset(ABC):
-    """Represents an preference optimization finetuning dataset."""
+    """Represents a preference optimization dataset."""
 
     @abstractmethod
     def create_reader(
@@ -115,7 +121,7 @@ npc = 10
 
 @final
 class GenericPreferenceOptimizationDataset(PreferenceOptimizationDataset):
-    """Represents a generic JSONL preferemce preference optimization dataset."""
+    """Represents a generic JSONL preference optimization dataset."""
 
     _files: Sequence[Path]
     _weights: Sequence[float]
@@ -138,88 +144,7 @@ class GenericPreferenceOptimizationDataset(PreferenceOptimizationDataset):
     @classmethod
     def from_path(cls, path: Path) -> GenericPreferenceOptimizationDataset:
         """Load a :class:`PreferenceOptimizationDataset` from ``path``."""
-        path = path.expanduser().resolve()
-
-        if not path.is_dir():
-            return GenericPreferenceOptimizationDataset(files=[path], weights=[1.0])
-
-        manifest_file = path.joinpath("MANIFEST")
-
-        try:
-            fp = manifest_file.open()
-        except FileNotFoundError:
-            fp = None
-        except OSError as ex:
-            raise RuntimeError(
-                f"{manifest_file} cannot be read. See nested exception for details."
-            ) from ex
-
-        # If the directory does not contain a MANIFEST file, treat all JSONL
-        # files as part of the dataset with equal weight.
-        if fp is None:
-            try:
-                files = list(path.glob("**/*.jsonl"))
-            except OSError as ex:
-                raise RuntimeError(
-                    f"The JSONL files under {path} cannot be retrieved. See nested exception for details."
-                ) from ex
-
-            weights = [1.0 for _ in range(len(files))]
-
-            return GenericPreferenceOptimizationDataset(files, weights=weights)
-
-        try:
-            content = list(fp)
-        except OSError as ex:
-            raise RuntimeError(
-                f"{manifest_file} cannot be read. See nested exception for details."
-            ) from ex
-        finally:
-            fp.close()
-
-        # Sort the JSONL files in alphabetical order.
-        content.sort()
-
-        files = []
-
-        weights = []
-
-        # Each line of the MANIFEST file corresponds to the path of a JSONL file
-        # and its weight (e.g. number of examples).
-        for idx, line in enumerate(content):
-
-            def raise_error() -> NoReturn:
-                raise DatasetError(
-                    f"Each line in {manifest_file} must represent a path to a JSONL file and a weight, but line {idx} is '{line}' instead."
-                )
-
-            fields = line.rstrip().split("\t")
-
-            if len(fields) != 2:
-                raise_error()
-
-            file_path = fields[0].strip()
-            if not file_path:
-                raise_error()
-
-            try:
-                file = path.joinpath(file_path)
-            except ValueError:
-                raise_error()
-
-            if not file.exists():
-                raise DatasetError(
-                    f"The file '{file}' referred at line {idx} in {manifest_file} does not exist."
-                )
-
-            files.append(file)
-
-            try:
-                weight = float(fields[1].strip())
-            except ValueError:
-                raise_error()
-
-            weights.append(weight)
+        files, weights = _load_files_and_weights(path)
 
         return GenericPreferenceOptimizationDataset(files, weights)
 
@@ -267,6 +192,9 @@ class GenericPreferenceOptimizationDataset(PreferenceOptimizationDataset):
 
         seed += 1
 
+        # Shard.
+        builder.shard(gang.rank, gang.size, allow_uneven=True)
+
         seed += gang.rank
 
         # Encode prompt and target texts.
@@ -277,21 +205,8 @@ class GenericPreferenceOptimizationDataset(PreferenceOptimizationDataset):
         builder.map(target_encoder, selector="tgt_chosen", num_parallel_calls=npc)
         builder.map(target_encoder, selector="tgt_rejected", num_parallel_calls=npc)
 
-        # Filter out long examples.
-        def skip(example: dict[str, Any]) -> bool:
-            chosen_len = len(example["src"]) + len(example["tgt_chosen"])
-            rejected_len = len(example["src"]) + len(example["tgt_rejected"])
-            return chosen_len <= max_seq_len and rejected_len <= max_seq_len
-
-        builder.filter(skip)
-
-        static_batching = isinstance(batching, StaticBatching)
-
-        # Shard.
-        builder.shard(gang.rank, gang.size, allow_uneven=not static_batching)
-
         def cat_source_and_target(example: dict[str, Any]) -> dict[str, Any]:
-            sample_id = example.get("id", None)
+            id_ = example.get("id", None)
 
             prompt_indices = example["src"]
             target_indices_chosen = example["tgt_chosen"]
@@ -300,19 +215,17 @@ class GenericPreferenceOptimizationDataset(PreferenceOptimizationDataset):
             indices_chosen = torch.cat([prompt_indices, target_indices_chosen])
             indices_rejected = torch.cat([prompt_indices, target_indices_rejected])
 
-            target_mask_chosen = torch.arange(len(indices_chosen)) >= len(
-                prompt_indices
-            )
-            target_mask_rejected = torch.arange(len(indices_rejected)) >= len(
-                prompt_indices
-            )
+            prompt_len = len(prompt_indices)
+
+            target_mask_chosen = torch.arange(len(indices_chosen)) >= prompt_len
+            target_mask_rejected = torch.arange(len(indices_rejected)) >= prompt_len
 
             return {
+                "id": id_,
                 "indices_chosen": indices_chosen,
                 "indices_rejected": indices_rejected,
                 "target_mask_chosen": target_mask_chosen,
                 "target_mask_rejected": target_mask_rejected,
-                "id": sample_id,
             }
 
         builder.map(cat_source_and_target, num_parallel_calls=npc)
@@ -327,10 +240,20 @@ class GenericPreferenceOptimizationDataset(PreferenceOptimizationDataset):
                 bucket_sizes,
                 selector="indices_chosen,indices_rejected",
                 skip_above_max_examples=True,
+                drop_remainder=drop_remainder,
             )
         else:
+            # Filter out long examples.
+            def skip(example: dict[str, Any]) -> bool:
+                chosen_len = len(example["indices_chosen"])
+                rejected_len = len(example["indices_rejected"])
+
+                return chosen_len <= max_seq_len and rejected_len <= max_seq_len
+
+            builder.filter(skip)
+
             # Bucket `batch_size` examples.
-            builder.bucket(batching.batch_size)
+            builder.bucket(batching.batch_size, drop_remainder=drop_remainder)
 
         # Shuffle buckets.
         if batch_shuffle_window != 1:
@@ -355,7 +278,7 @@ class GenericPreferenceOptimizationDataset(PreferenceOptimizationDataset):
         # Prefetch `num_prefetch` batches in background.
         builder.prefetch(num_prefetch)
 
-        # Wrap examples with `SequenceBatch`.
+        # Wrap examples with `PreferenceOptimizationBatch`.
         def to_batch(example: dict[str, Any]) -> PreferenceOptimizationBatch:
             indices_chosen = cast(SequenceData, example["indices_chosen"])
             indices_rejected = cast(SequenceData, example["indices_rejected"])
@@ -368,13 +291,15 @@ class GenericPreferenceOptimizationDataset(PreferenceOptimizationDataset):
             )
 
             target_mask_chosen = example["target_mask_chosen"]["seqs"].to(gang.device)
-            target_mask_rejected = example["target_mask_rejected"]["seqs"].to(
-                gang.device
-            )
+            target_mask_rejected = example["target_mask_rejected"]["seqs"].to(gang.device)  # fmt: skip
 
             batch_chosen = SequenceBatch(
-                seqs_chosen, padding_mask_chosen, target_mask_chosen, example=example
+                seqs_chosen,
+                padding_mask_chosen,
+                target_mask_chosen,
+                example=example,
             )
+
             batch_rejected = SequenceBatch(
                 seqs_rejected,
                 padding_mask_rejected,
@@ -382,9 +307,7 @@ class GenericPreferenceOptimizationDataset(PreferenceOptimizationDataset):
                 example=example,
             )
 
-            return PreferenceOptimizationBatch(
-                chosen=batch_chosen, rejected=batch_rejected
-            )
+            return PreferenceOptimizationBatch(batch_chosen, batch_rejected)
 
         pipeline = builder.map(to_batch).and_return()
